fix(2): execute commands piped with spaces SST-54
It looks like input coming from 'echo "command" | ./hsh' introduces
a NULL pointer that has to be handled by the trim function. When
the same input comes from the terminal, there is no NULL pointer
that needs to be handled. trim( ) was not handling NULL pointers so
it attempted to dereference the NULL pointer and caused errors when
input was piped but not when input was from the terminal.

I don't understand why there is a difference. I will look into that
later.

See SST-54 (I am testing the autoreferencing feature of GitHub).