refactor: define Argument struct
I am introducing a new type, Argument, as a first step in making
the code more readable and object-oriented. It will store an array
of strings (arguments) and a count of those arguments.

I am imagining it will have the following responsibilities:

	- initialise a character array of arguments from a string
	- deinitialise its memory allocations

Although I haven't worked out the details, I am expecting Argument
to be used by a Command or Execute object to launch a program.