refactor: introduce arguments Parameter Object
This continues the attempt to refactor the codebase to an object-
oriented style.

I want functions to receive objects instead of unencapsulated data.
This will make interactions clearer and create a better separation
of concerns with increased cohesion. The command execution logic
was receiving a basic type as input; it continues to receive a
basic type but it is now getting that data via the member of an
Argument object.

The next step is to change the interface of the method so that it
takes an Argument object directly.