refactor: add Command type
This is the second step in creating a more object-oriented shell.
Previously, an Arguments object was introduced. It is now used in
the call that processes commands.

The next phase is to encapsulate program execution logic into a
Command structure. One of its members will be the previously
introduced Arguments object.